Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
python-dotenv
cloudinary
psycopg2-binary
Pillow